fix(ci): maintainability gate — move TestGrokAdapter into test_grok_adapter to keep test_adapters.py under the 1500-line limit
